Rename quickstart -> templating
